feat: Move PayType enum to enum package
